Added new output data format 'protobuf'.
It generates protocol buffer version 3 formatted binaries.
Message format definition could be found from metric/proto/metric.proto.

Output formats like `file` requires a mechanism to identify the beginning and the end point of each message. In Protocol Buffers, there is no way to do it by the format itself. To solve this, you can set the optional parameter `prepend_length` to `true`. If this parameter is set, the length of each message will be prepended in front of real message.